# application/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: Mechanic API
  description: |
    A comprehensive REST API for managing auto repair shop operations including mechanics, customers, inventory, and service tickets.
    
    ## Authentication
    This API uses JWT (JSON Web Tokens) for authentication. After logging in, include the token in the Authorization header:
    ```
    Authorization: Bearer <your_jwt_token>
    ```
    
    ## Rate Limiting
    - Customer endpoints: 100 requests per hour
    - Login endpoints: 10 requests per minute
    - Specific endpoints may have additional rate limits
    
    ## Error Handling
    The API uses standard HTTP status codes and returns error messages in JSON format.
  version: 1.0.0
  contact:
    name: API Support
    email: support@mechanicapi.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.mechanic.com/v1
    description: Production server
...
paths:
...
  /inventory/:
    get:
      tags:
        - Inventory
      summary: Get all inventory items
      description: Retrieve a list of all inventory items
      responses:
        '200':
          description: Inventory items retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InventoryItem'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - Inventory
      summary: Create a new inventory item
      description: Add a new item to the inventory (requires mechanic authentication)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InventoryCreate'
      responses:
        '201':
          description: Inventory item created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryItem'
        '400':
          description: Validation errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /inventory/{item_id}:
    get:
      tags:
        - Inventory
      summary: Get specific inventory item by ID
      description: Retrieve a specific inventory item by its ID
      parameters:
        - name: item_id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Inventory item retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryItem'
        '404':
          description: Inventory item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Inventory
      summary: Update inventory item
      description: Update a specific inventory item by its ID (requires mechanic authentication)
      security:
        - BearerAuth: []
      parameters:
        - name: item_id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InventoryUpdate'
      responses:
        '200':
          description: Inventory item updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryItem'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Inventory item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Inventory
      summary: Delete inventory item
      description: Delete a specific inventory item by its ID (requires mechanic authentication)
      security:
        - BearerAuth: []
      parameters:
        - name: item_id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '204':
          description: Inventory item deleted successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Inventory item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /inventory/low-stock:
    get:
      tags:
        - Inventory
      summary: Get inventory items with low stock
      description: Retrieve inventory items that are below their minimum stock level
      parameters:
        - name: threshold
          in: query
          schema:
            type: integer
            description: Custom low stock threshold (defaults to min_stock_level)
            example: 10
      responses:
        '200':
          description: Low stock items retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InventoryItem'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /inventory/category/{category_name}:
    get:
      tags:
        - Inventory
      summary: Get inventory items by category
      description: Retrieve inventory items filtered by category
      parameters:
        - name: category_name
          in: path
          required: true
          schema:
            type: string
            example: "Lubricants"
      responses:
        '200':
          description: Category items retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InventoryItem'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Service Ticket Routes
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoints
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Registration failed"
        message:
          type: string
          example: "Internal server error occurred"
        details:
          type: string
          example: "Database connection error"
    ValidationError:
      type: object
      properties:
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          example:
            email: ["Not a valid email address"]
            password: ["Password must be at least 8 characters"]

    # Authentication Schemas
...
    InventoryItem:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Engine Oil"
        description:
          type: string
          example: "Synthetic engine oil 5W-30"
        price:
          type: number
          format: float
          example: 29.99
        quantity:
          type: integer
          example: 50
        category:
          type: string
          example: "Lubricants"
        min_stock_level:
          type: integer
          example: 10
        supplier:
          type: string
          example: "AutoParts Inc."
        created_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2024-01-15T14:45:00Z"
    InventoryCreate:
      type: object
      required:
        - item_name
      properties:
        item_name:
          type: string
          example: "Engine Oil"
        price:
          type: number
          format: float
          example: 29.99
        quantity:
          type: integer
          example: 50
        description:
          type: string
          example: "Synthetic engine oil 5W-30"
        category:
          type: string
          example: "Lubricants"
        min_stock_level:
          type: integer
          example: 10
        supplier:
          type: string
          example: "AutoParts Inc."
    InventoryUpdate:
      type: object
      properties:
        item_name:
          type: string
          example: "Premium Engine Oil"
        price:
          type: number
          format: float
          example: 34.99
        quantity:
          type: integer
          example: 45
        description:
          type: string
          example: "Premium synthetic engine oil 5W-30"
        category:
          type: string
          example: "Lubricants"
        min_stock_level:
          type: integer
          example: 15
        supplier:
          type: string
          example: "Premium AutoParts Inc."

    # Service Ticket Schemas
...